third_party/blink/web_tests/webaudio/codec-tests/webm/webm-decode.html

<!DOCTYPE html>
<html>

<head>
  <title>
    WebM decoding test
  </title>
  <script src="../../../resources/testharness.js"></script>
  <script src="../../../resources/testharnessreport.js"></script>
  <script src="../../resources/audit-util.js"></script>
  <script src="../../resources/audio-file-utils.js"></script>
  <script src="../../resources/buffer-loader.js"></script>
  <script src="../resources/codec-tests.js"></script>
</head>

<body>
  <script>
    runCodecTest({
      encodedFileName: 'test-webm.webm',
      // Absolute tolerance for sample-level comparison.
      absTolerance: 4.0234e-5,
      // Use 48 kHz to avoid resampling in decodeAudioData.
      sampleRate: 48000,
      // Minimum acceptable signal-to-noise ratio (SNR) in decibels for each
      // channel.
      minSnrDb: 110.97,
      description: 'Decoded WebM matches reference WAV within tolerance',
    });
  </script>
</body>

</html>
